Selectmenu: Move selectmenu into widgets folder
