+++ tests/unit/selectmenu/options.js
@@ -1,6 +1,6 @@
 define( [
 	"jquery",
-	"ui/selectmenu"
+	"ui/widgets/selectmenu"
 ], function( $ ) {
 
 module( "selectmenu: options" );
